Note model changes

+++ routes/listRoutes.js
@@ -9,13 +9,15 @@ router.all('*', cors());
 
 // Adding a new note
 router.post('/addNote', authorize, asyncMiddleware(async (req, res) => {
+    let userId = req.body.userId;
     let title = req.body.title;
     let body = req.body.body;
-    let addedNote = await saveNote(title, body);
+    let modifiedTime = req.body.modifiedTime;
+    let addedNote = await saveNote(userId, title, body, modifiedTime);
     if (addedNote) {
         res.send({
             id: addedNote._id,
-            status: 'Note added successfully...'
+            status: 'Note added successfully...',
         });
     } else {
         res.send({ 
@@ -25,11 +27,13 @@ router.post('/addNote', authorize, asyncMiddleware(async (req, res) => {
 }));
 
 // Saving a note
-async function saveNote(title, body) {
+async function saveNote(userId, title, body, modifiedTime) {
     try {
         const note = new Note({
+            userId,
             title,
-            body
+            body,
+            lastModified: modifiedTime
         });
         const result = await note.save();
         console.log(result);
@@ -49,16 +53,18 @@ router.get('/getNote/:id', authorize, asyncMiddleware(async (req, res) => {
 
 // Get all notes
 router.get('/getAllNotes', authorize, asyncMiddleware(async (req, res) => {
-    const notes = await Note.find().select('title body dateCreated lastModified');
+    let userId = req.header('x-user-id');
+    const notes = await Note.find({ userId: userId }).select('title body lastModified');
     res.send(notes);
 }));
 
 // Update a note by id
 router.post('/updateNote/:id', authorize, asyncMiddleware(async (req, res) => {
+    let userId = req.body.userId;
     let newTitle = req.body.title;
     let newBody = req.body.body;
     let id = req.params.id;
-    await Note.findByIdAndUpdate({_id: id}, {
+    await Note.findByIdAndUpdate({ userId: userId, _id: id }, {
         $set: {
             title: newTitle,
             body: newBody,
@@ -70,8 +76,9 @@ router.post('/updateNote/:id', authorize, asyncMiddleware(async (req, res) => {
 
 // Remove a note by id
 router.delete('/deleteNote/:id', authorize, asyncMiddleware(async (req, res) => {
+    let userId = req.body.userId;
     let id = req.params.id;
-    await Note.findByIdAndDelete({_id: id});
+    await Note.findByIdAndDelete({ userId: userId,  _id: id });
     res.send('Note deleted successfully');
 }));
